open-100: Open Sample Model.  Add Momentum.ifc test fixture.  Refactor Open* components, one per file.  profile-100: use nested context and helpers consistently.  Start a cleanup of helpers, sorting with more docs, spacing lint.
